+++ client/src/Pages/EmployeeDash/BookingDetails.js
@@ -1,11 +1,11 @@
 import React from "react";
-import BookingForm from "./BookingForm";
-import Table from "./Table";
+import BookingForm from "./Form/BookingForm";
+import Table from "./Table/Table";
 import "./EmployeeDsh.css";
 
 const BookingDetails = ({ isSplitView }) => {
 	return (
-		<div class="w-100 h-75 d-flex justify-content-around">
+		<div class="w-100 d-flex justify-content-around">
 			{isSplitView ? (
 				<>
 					<BookingForm />
