+++ package.json
@@ -54,8 +54,10 @@
     "prettier": "^3.1.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
+    "react-datepicker": "^4.25.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.20.1",
+    "react-toastify": "^9.1.3",
     "style-loader": "^3.3.3",
     "wait-on": "^7.2.0",
     "webpack": "^5.89.0",
